fixed integer overflow for tables with more than 21 million rows. extended support to 92 000 000 billion rows

+++ dbptk-core/src/main/java/com/databasepreservation/modules/siard/in/content/SIARD1ContentImportStrategy.java
@@ -76,7 +76,7 @@ public class SIARD1ContentImportStrategy extends DefaultHandler implements Conte
   private BinaryCell currentBlobCell;
   private SimpleCell currentClobCell;
   private Row row;
-  private int rowIndex;
+  private long rowIndex;
   private long currentTableTotalRows;
 
   private long lastProgressTimestamp;
