+++ dbptk-core/src/main/java/com/databasepreservation/modules/siard/in/content/SIARD2ContentImportStrategy.java
@@ -77,7 +77,7 @@ public class SIARD2ContentImportStrategy extends DefaultHandler implements Conte
   private BinaryCell currentBlobCell;
   private SimpleCell currentClobCell;
   private Row row;
-  private int rowIndex;
+  private long rowIndex;
   private long currentTableTotalRows;
 
   private long lastProgressTimestamp;
